refactor: renamed it to RasterLegend
