+++ sites/geohub/src/components/controls/RasterLegend.svelte
